Changes made for Bug # 17326
In file download the Category Record Id for that particular category is passed as a parameter in the request.
After some of the changes done in DE, this record was not getting populated while populating the File upload attribute in edit mode.

The main cause was that, the child container's data value map was not getting populated properly and was casuing this problem.
As per now, child container's data value map also gets populated.

Code Reviewed By : Kunal

Code Reviewed On : June 08 2010

SVN-Revision: 26979